api - refactor structure to split routes and lib, add deta dep
