+++ modules/lib/products.js
