+++ modules/lib/util.js
