+++ modules/routes/notes.js
@@ -1,7 +1,7 @@
 const axios = require('axios')
 const { WebClient } = require('@slack/web-api')
-const { lambda, getFuncName } = require('./util')
 
+const { lambda, getFuncName } = require('../lib/util')
 
 const { SLACK_OAUTH, DEPLOYED } = process.env
 const web = new WebClient(SLACK_OAUTH)
