+++ package.json
@@ -15,6 +15,7 @@
     "@eqworks/release": "^3.1.0",
     "@slack/web-api": "^5.7.0",
     "axios": "^0.21.1",
+    "deta": "^1.0.1",
     "express": "^4.16.3",
     "googleapis": "^66.0.0",
     "lodash.samplesize": "^4.2.0",
